SVG support
This is a very simple solution but it is mostly applicable.
It is aimed not to decrease the speed of runtime mode.

SVG elements creation is followed by svgPosition variable in
compiler.ts.
Only when inside SVG elements (svgPosition has a value) that uses
createSvgElement instead of createElement and setAttribute instead of
"." Otherwise if the svgPosition is null normal process continues in the
compiler.